+++ .ci-scripts/x86-64-unknown-linux-musl-release.bash
@@ -39,7 +39,7 @@ ASSET_DESCRIPTION="https://github.com/ponylang/ponyc"
 # Build pony installation
 echo "Building ponyc installation..."
 make install prefix=${BUILD_PREFIX} default_pic=${PIC} arch=${ARCH} \
-  -j${MAKE_PARALLELISM} -f Makefile-lib-llvm symlink=no use=llvm_link_static
+  link=static -j${MAKE_PARALLELISM} -f Makefile-lib-llvm symlink=no
 
 # Package it all up
 echo "Creating .tar.gz of ponyc installation..."